+++ src/containers/bookmark/type.ts
@@ -1,0 +1,9 @@
+interface Bookmark {
+    user_id: any;
+    id: number;
+    shopname: string;
+    location: string;
+    type: string;
+    lat: number;
+    lng: number;
+  }
