--- src/containers/map/index.tsx
@@ -1,1 +1,0 @@
-1
